Move serviceProvisions and sibling services to Service class
