+++ src/components/ServicePane.js
@@ -12,7 +12,6 @@ import Eligibility from "../components/Eligibility";
 import OpeningTimes from "../components/OpeningTimes";
 import colors from "../constants/theme";
 import fixtures from "../../fixtures/services";
-import serviceProvisions from "../constants/service-provisions";
 
 var palette = colors.getPalette();
 
@@ -26,30 +25,6 @@ export default class ServicePane extends React.Component {
         return this.props.service.description.split('.', 1)[0] + '.';
     }
 
-    /**
-     * serviceProvisions:
-     *
-     * An array of things this service provides built using a bucket-of-words
-     * approach from the service's full description */
-    /* flow:disable */
-    get serviceProvisions(): Array<string> {
-        if (this._serviceProvisions) {
-            return this._serviceProvisions;
-        }
-
-        var service = this.props.service;
-        this._serviceProvisions = [];
-
-        for (var provision of serviceProvisions) {
-            var forms = [provision.cname].concat(provision.forms || []);
-
-            if (_.some(forms.map(form => new RegExp(form, 'i')),
-                       form => service.description.match(form))) {
-                this._serviceProvisions.push(provision.cname);
-            }
-        }
-
-        return this._serviceProvisions;
     }
 
     render(): React.Element {
@@ -83,7 +58,7 @@ export default class ServicePane extends React.Component {
 
                 <h2>What you can get here</h2>
                 <ul>
-                    {this.serviceProvisions.map(
+                    {object.serviceProvisions.map(
                         (provision, index) => <li key={index}>{provision}</li>
                     )}
                 </ul>
